cloud/backend/requirements.txt
Django
psycopg2-binary
gunicorn
django-environ
djangorestframework
django-debug-toolbar
pdf2image
django-cors-headers
celery
elasticsearch-dsl
django-elasticsearch-dsl
django-elasticsearch-dsl-drf
djangorestframework-bulk
django-model-utils